DAOS-12392 test: Allow test_core_files to pass in launch.py
Quick-build: true
Skip-unit-tests: true
Test-tag: pr test_core_files

Required-githooks: true
